move scripts, re-generate from latest docs
