+++ scripts/docs/cli_generate_spec.py
